Clean restatapi cache

Remove all objects from the .restatapi_env except the configuration file, API version number, download method and the country codes. In addition, it deletes all the .rds files from the default and selected cache directory. See get_eurostat_data for more on cache.

clean_restatapi_cache(cache_dir = NULL, verbose = FALSE)

Arguments

cache_dir

a path to cache directory. If NULL (default) it will clean default temporary cache directory (file.path(tempdir(),"restatapi")). The default cache directory is used when the provided cache_dir does not exist. Directory can also be set with options(restatapi_cache_dir=...).

verbose

a logical value with default FALSE, so detailed messages (for debugging) will not be printed. Can be set also with options(restatapi_verbose=TRUE)

Examples

clean_restatapi_cache(verbose=TRUE) 
#> 
#> clean_restatapi_cache - All objects except from 'cfg', 'rav', 'cc' and 'dmethod' are removed from '.restatapi_env'.