Extract data values from SDMX XML

Extracts the data values from the SDMX XML data file

extract_data(xml_lf, keep_flags = FALSE,
  stringsAsFactors = default.stringsAsFactors())

Arguments

xml_lf
an XML leaf with data series from an SDMX XML file

keep_flags
a boolean if to extract the observation status (flag) information from the XML file. The default value is FALSE

stringsAsFactors
if TRUE (the default) the columns are converted to factors. If FALSE they are returned as a character.

Value

a data frame with

Details

It is a subfunction to use in the get_eurostat_data function.

Examples

id<-"agr_r_milkpr"
toc<-get_eurostat_toc()
bulk_url<-toc$downloadLink.sdmx[toc$code==id]
temp<-tempfile()
download.file(bulk_url,temp)
sdmx_xml<-xml2::read_xml(unzip(temp, paste0(id,".sdmx.xml")))
xml_leafs<-xml2::xml_find_all(sdmx_xml,".//data:Series")
extract_data(xml_leafs[1])

#>    FREQ milkitem dairyprod geo TIME_FORMAT TIME_PERIOD OBS_VALUE
#> 1     A      PRO    D1110A  AT         P1Y        1996    2956.6
#> 2     A      PRO    D1110A  AT         P1Y        1997      3015
#> 3     A      PRO    D1110A  AT         P1Y        1998    3042.6
#> 4     A      PRO    D1110A  AT         P1Y        1999    3131.9
#> 5     A      PRO    D1110A  AT         P1Y        2000   3233.20
#> 6     A      PRO    D1110A  AT         P1Y        2001   3299.60
#> 7     A      PRO    D1110A  AT         P1Y        2002   3292.20
#> 8     A      PRO    D1110A  AT         P1Y        2003   3229.90
#> 9     A      PRO    D1110A  AT         P1Y        2004   3137.30
#> 10    A      PRO    D1110A  AT         P1Y        2005   3113.70
#> 11    A      PRO    D1110A  AT         P1Y        2006   3146.66
#> 12    A      PRO    D1110A  AT         P1Y        2007   3155.07
#> 13    A      PRO    D1110A  AT         P1Y        2008   3195.95
#> 14    A      PRO    D1110A  AT         P1Y        2009   3229.81
#> 15    A      PRO    D1110A  AT         P1Y        2010   3257.74
#> 16    A      PRO    D1110A  AT         P1Y        2011   3307.13
#> 17    A      PRO    D1110A  AT         P1Y        2012   3382.00
#> 18    A      PRO    D1110A  AT         P1Y        2013   3393.06
#> 19    A      PRO    D1110A  AT         P1Y        2014   3493.86
#> 20    A      PRO    D1110A  AT         P1Y        2015   3537.76
#> 21    A      PRO    D1110A  AT         P1Y        2016   3627.61
#> 22    A      PRO    D1110A  AT         P1Y        2017   3712.73