Filter raw data locally

Filter downloaded full raw dataset on local computer if the get_eurostat_data has not provided data due to too large datasets for the REST API.

filter_raw_data(raw_data = NULL, filter_table = NULL, date_filter = FALSE)

Arguments

raw_data

an input data.table dataset resulted from the call of the get_eurostat_raw function

filter_table

a data table with values for the concepts or time to be filtered out which can be generated by the create_filter_table function

date_filter

a logical value. If TRUE the filter table should be applied to the time columns of the raw_data. The default is FALSE, in this case the filters applied to the other columns of the raw_data.

Value

a filtered data.table containing only the rows of raw_data which fulfills the conditions in the filter_table

Details

It is a sub-function to use in the get_eurostat_data to filter data on the local computer if the direct response from REST API did not provide data because of too large data set (more than 30 thousands observations). The filter_table contains always at least two columns. In case if date_filter=TRUE then the two columns should have the following names and the provided conditions are applied to the time column of the the raw_data data.table.

sd	Starting date to be included, where date is formatted as yyyy[-mm][-dd] (the month and day are optional)

ed	End date of the period to be included in the dataset formatted as yyyy[-mm][-dd] (the month and day are optional)

In case if date_filter=FALSE then the columns should have the following names:

concept	Containing concept names, which is a column name in the raw_data data.table

code	A possible code under the given concept, which is a value in the column of the raw_data data.table defined by the concept

See also

Examples

# \dontshow{

# }
# \donttest{
id<-"tus_00age"
options(timeout=2)
rd<-get_eurostat_raw(id)
dsd<-get_eurostat_dsd(id)
ft<-create_filter_table(c("TIME_SP","Hungary",'T'),FALSE,dsd)
filter_raw_data(rd,ft)
#>      freq    unit sex    age acl00 geo time values
#>   1:    A TIME_SP   T  TOTAL   AC0  HU 2010  11:32
#>   2:    A TIME_SP   T  TOTAL  AC01  HU 2010   8:34
#>   3:    A TIME_SP   T  TOTAL  AC02  HU 2010   1:42
#>   4:    A TIME_SP   T  TOTAL  AC03  HU 2010   1:16
#>   5:    A TIME_SP   T  TOTAL  AC1A  HU 2010   2:23
#>  ---                                              
#> 388:    A TIME_SP   T Y_GE65  AC9A  HU 2010   0:36
#> 389:    A TIME_SP   T Y_GE65  AC9B  HU 2010   0:00
#> 390:    A TIME_SP   T Y_GE65  AC9C  HU 2010   0:01
#> 391:    A TIME_SP   T Y_GE65  AC9D  HU 2010   0:09
#> 392:    A TIME_SP   T Y_GE65 TOTAL  HU 2010  24:00
options(timeout=60)
# }