Get Eurostat data in a standardized format

Download data sets from Eurostat database and put in a standardized format.

get_eurostat_bulk(id, cache = TRUE, update_cache = FALSE,
  cache_dir = NULL, compress_file = TRUE,
  stringsAsFactors = default.stringsAsFactors(), select_freq = NULL,
  keep_flags = FALSE, verbose = FALSE, ...)

Arguments

id
A code name for the dataset of interest. See search_eurostat_toc for details how to get an id.

cache
a logical whether to do caching. Default is TRUE.

update_cache
a logical with a default value FALSE, whether to update cache. Can be set also with options(restatapi_update=TRUE)

cache_dir
a path to a cache directory. The NULL (default) uses the memory as cache. If the folder if the cache_dir directory does not exist it saves in the 'restatapi' directory under the temporary directory from tempdir(). Directory can also be set with option(restatapi_cache_dir=...).

compress_file
a logical whether to compress the RDS-file in caching. Default is TRUE.

stringsAsFactors
if TRUE (the default) variables are not numeric then they are converted to factors. If the value FALSE they are returned as a characters.

select_freq
a character symbol for a time frequency when a dataset has multiple time frequencies. Possible values are: A = annual, S = semi-annual, Q = quarterly, M = monthly. The default is NULL as most datasets have just one time frequency and in this case if there are multiple frequencies, then only the most common frequency kept. If all the frequencies needed the get_eurostat_raw can be used.

keep_flags
a logical whether the observation status (flags) - e.g. "confidential", "provisional", etc. - should be kept in a separate column or if they can be removed. Default is FALSE. For flag values see: http://ec.europa.eu/eurostat/data/database/information.

verbose
A boolean with default FALSE, so detailed messages (for debugging) will not printed. Can be set also with options(restatapi_verbose=TRUE)

...
parameter to pass on the load_cfg function

Value

a data.table. One column for each dimension in the data, the time column for a time dimension, the values column for numerical values and the flags column if the keep_flags=TRUE. Eurostat data does not include all missing values. The missing values are dropped if all dimensions are missing on particular time.

Details

Data sets are downloaded from the Eurostat bulk download facility in SDMX format and filtered for a unique time frequency. If no frequency is selected and there are multiple frequencies in the dataset, then the most common value is used.

Compared to get_eurostat_raw the frequency column is removed and the original column names for the time period, observation values and status are renamed to "time", "values" and "flags".

By default all datasets cached as they are often rather large. The datasets cached in memory (default) or can be stored in a temporary directory if cache_dir or option(restatpi_cache_dir) is defined. The cache can be emptied with clean_restatapi_cache.

The id, is a value from the code column of the table of contents (get_eurostat_toc), and can be searched for with the search_eurostat_toc function. The id value can be retrieved from the Eurostat database as well. The Eurostat database gives codes in the Data Navigation Tree after every dataset in parenthesis.

See also

Examples

dt<-get_eurostat_bulk("agr_r_milkpr",keep_flags=TRUE)
options(restatapi_update=TRUE)
dt<-get_eurostat_bulk("avia_par_ee")

#> There are multiple frequencies in the dataset. The 'M' is selected as it is the most common frequency.

dt<-get_eurostat_bulk("avia_par_ee",select_freq="A",verbose=TRUE)

#> https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&downfile=table_of_contents.xml

#> The TOC was cached in memory ('toc.xml.en' in '.restatapi_env'). The previous value was overwritten.

#> https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&downfile=table_of_contents.xml

#> The TOC was cached in memory ('toc.xml.en' in '.restatapi_env'). The previous value was overwritten.

#> 6263bulk url: http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/avia_par_ee.sdmx.zip

#> The bulk data was cached in memory ('b_avia_par_ee-27.03.2019-0-A' in '.restatapi_env').

options(restatapi_update=FALSE)
dt<-get_eurostat_bulk("agr_r_milkpr",cache_dir=tempdir(),compress_file=FALSE,verbose=TRUE)

#> The 'toc.xml.en' was loaded from 'toc.xml.en' in '.restatapi_env'.

#> The TOC was cached previously in memory ('toc.xml.en' in '.restatapi_env'). It remained unchanged.

#> The 'b_agr_r_milkpr-29.03.2019-0' was loaded from 'b_agr_r_milkpr-29.03.2019-1' in '.restatapi_env'.

#> The bulk data was cached in the file /tmp/Rtmp012GKw/b_agr_r_milkpr-29.03.2019-0.rds.