Extract values from XML

Extracts the values from the XML Data Structure Definition (DSD) file

extract_dsd(concept = NULL, dsd_xml = NULL)

Arguments

concept
a character vector with concept ids

dsd_xml
an XML file with DSD content

Value

a matrix with 3 columns if the concepts has code list in the DSD file

Details

It is a subfunction to use in the get_eurostat_dsd function.

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}
# }
dsd_url<-"http://ec.europa.eu/eurostat/SDMX/diss-web/rest/datastructure/ESTAT/DSD_nama_10_a10_e"
tryCatch({
  dsd_xml<-xml2::read_xml(dsd_url)},
  error=function(e){
  message("Unable to download the xml file.\n",e)},
  warning=function(w){
  message("Unable to download the xml file.\n",w)})
if (exists("dsd_xml")) {extract_dsd("GEO",dsd_xml)}

#>       concept xml_clc    
#>  [1,] "GEO"   "EU28"     
#>  [2,] "GEO"   "EU27_2019"
#>  [3,] "GEO"   "EU15"     
#>  [4,] "GEO"   "EA"       
#>  [5,] "GEO"   "EA19"     
#>  [6,] "GEO"   "EA12"     
#>  [7,] "GEO"   "BE"       
#>  [8,] "GEO"   "BG"       
#>  [9,] "GEO"   "CZ"       
#> [10,] "GEO"   "DK"       
#> [11,] "GEO"   "DE"       
#> [12,] "GEO"   "EE"       
#> [13,] "GEO"   "IE"       
#> [14,] "GEO"   "EL"       
#> [15,] "GEO"   "ES"       
#> [16,] "GEO"   "FR"       
#> [17,] "GEO"   "HR"       
#> [18,] "GEO"   "IT"       
#> [19,] "GEO"   "CY"       
#> [20,] "GEO"   "LV"       
#> [21,] "GEO"   "LT"       
#> [22,] "GEO"   "LU"       
#> [23,] "GEO"   "HU"       
#> [24,] "GEO"   "MT"       
#> [25,] "GEO"   "NL"       
#> [26,] "GEO"   "AT"       
#> [27,] "GEO"   "PL"       
#> [28,] "GEO"   "PT"       
#> [29,] "GEO"   "RO"       
#> [30,] "GEO"   "SI"       
#> [31,] "GEO"   "SK"       
#> [32,] "GEO"   "FI"       
#> [33,] "GEO"   "SE"       
#> [34,] "GEO"   "UK"       
#> [35,] "GEO"   "IS"       
#> [36,] "GEO"   "LI"       
#> [37,] "GEO"   "NO"       
#> [38,] "GEO"   "CH"       
#> [39,] "GEO"   "ME"       
#> [40,] "GEO"   "MK"       
#>       xml_cln                                                                                        
#>  [1,] "European Union - 28 countries"                                                                
#>  [2,] "European Union - 27 countries (from 2019)"                                                    
#>  [3,] "European Union - 15 countries (1995-2004)"                                                    
#>  [4,] "Euro area (EA11-2000, EA12-2006, EA13-2007, EA15-2008, EA16-2010, EA17-2013, EA18-2014, EA19)"
#>  [5,] "Euro area (19 countries)"                                                                     
#>  [6,] "Euro area (12 countries)"                                                                     
#>  [7,] "Belgium"                                                                                      
#>  [8,] "Bulgaria"                                                                                     
#>  [9,] "Czechia"                                                                                      
#> [10,] "Denmark"                                                                                      
#> [11,] "Germany (until 1990 former territory of the FRG)"                                             
#> [12,] "Estonia"                                                                                      
#> [13,] "Ireland"                                                                                      
#> [14,] "Greece"                                                                                       
#> [15,] "Spain"                                                                                        
#> [16,] "France"                                                                                       
#> [17,] "Croatia"                                                                                      
#> [18,] "Italy"                                                                                        
#> [19,] "Cyprus"                                                                                       
#> [20,] "Latvia"                                                                                       
#> [21,] "Lithuania"                                                                                    
#> [22,] "Luxembourg"                                                                                   
#> [23,] "Hungary"                                                                                      
#> [24,] "Malta"                                                                                        
#> [25,] "Netherlands"                                                                                  
#> [26,] "Austria"                                                                                      
#> [27,] "Poland"                                                                                       
#> [28,] "Portugal"                                                                                     
#> [29,] "Romania"                                                                                      
#> [30,] "Slovenia"                                                                                     
#> [31,] "Slovakia"                                                                                     
#> [32,] "Finland"                                                                                      
#> [33,] "Sweden"                                                                                       
#> [34,] "United Kingdom"                                                                               
#> [35,] "Iceland"                                                                                      
#> [36,] "Liechtenstein"                                                                                
#> [37,] "Norway"                                                                                       
#> [38,] "Switzerland"                                                                                  
#> [39,] "Montenegro"                                                                                   
#> [40,] "North Macedonia"