Extract the Data Structure Definition content from SDMX XML

Extracts values from the XML Data Structure Definition (DSD) file

extract_dsd(concept = NULL, dsd_xml = NULL)

Arguments

concept
a character vector with a concept id

dsd_xml
an XML file with DSD content

Value

a matrix with 3 columns if the provided concept has a code list in the DSD file. The first column is the provided concept. The second column is the possible codes under the given concept. The last column is the name/description for the code in the second column, which can be used as labels.

Details

It is a sub-function to use in the get_eurostat_dsd function.

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}    
# }
# \donttest{
dsd_url<-"https://ec.europa.eu/eurostat/SDMX/diss-web/rest/datastructure/ESTAT/DSD_med_rd6"
tryCatch({
  dsd_xml<-xml2::read_xml(dsd_url)}, 
  error=function(e){
  message("Unable to download the xml file.\n",e)}, 
  warning=function(w){
  message("Unable to download the xml file.\n",w)}) 
if (exists("dsd_xml")) {extract_dsd("GEO",dsd_xml)} 

#>      concept xml_clc xml_cln  
#> [1,] "GEO"   "MA"    "Morocco"
#> [2,] "GEO"   "JO"    "Jordan" 

# }