Extract values from XML

Extracts the values of a node from the Eurostat XML Table of contents (TOC) file

extract_toc(ns)

Arguments

ns
An XML node set from the XML TOC file

Value

a character vector with all the values of the node set.

Details

It is a sub-function to use in the get_eurostat_toc function.

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}
# }
cfg<-get("cfg",envir=.restatapi_env)
rav<-get("rav",envir=.restatapi_env)
toc_endpoint<-eval(parse(text=paste0("cfg$TOC_ENDPOINT$'",rav,"'$ESTAT$xml")))
# \donttest{
xml_leafs<-xml2::xml_find_all(xml2::read_xml(toc_endpoint),".//nt:leaf")
extract_toc(xml_leafs[1])

#>  [1] "Consumers - monthly data"                                                                                       
#>  [2] "Consommateurs - données mensuelles"                                                                             
#>  [3] "Verbraucher - monatliche Daten"                                                                                 
#>  [4] "ei_bsco_m"                                                                                                      
#>  [5] "02.03.2020"                                                                                                     
#>  [6] "02.03.2020"                                                                                                     
#>  [7] "1980M01"                                                                                                        
#>  [8] "2020M02"                                                                                                        
#>  [9] "268082"                                                                                                         
#> [10] ""                                                                                                               
#> [11] ""                                                                                                               
#> [12] ""                                                                                                               
#> [13] ""                                                                                                               
#> [14] ""                                                                                                               
#> [15] ""                                                                                                               
#> [16] "https://ec.europa.eu/eurostat/cache/metadata/en/ei_bcs_esms.htm"                                                
#> [17] "https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bcs_esms.sdmx.zip"
#> [18] "https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bsco_m.tsv.gz"        
#> [19] "https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bsco_m.sdmx.zip"      

# }