Extract the text of the table of contents from SDMX XML

Extracts the values of a node from the Eurostat XML Table of contents (TOC) file

extract_toc(ns)

Arguments

ns
an XML node set from the XML TOC file

Value

a character vector with all the values of the node set.

Details

It is a sub-function to use in the get_eurostat_toc function.

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}    
# }
cfg<-get("cfg",envir=restatapi::.restatapi_env) 
rav<-get("rav",envir=restatapi::.restatapi_env)
toc_endpoint<-eval(parse(text=paste0("cfg$TOC_ENDPOINT$'",rav,"'$ESTAT$xml")))
# \donttest{
options(timeout=2)
tryCatch(xml_leafs<-xml2::xml_find_all(xml2::read_xml(toc_endpoint),".//nt:leaf"))
restatapi::extract_toc(xml_leafs[1])

#>  [1] "Consumers - monthly data"                                                                                       
#>  [2] "Consommateurs - données mensuelles"                                                                             
#>  [3] "Verbraucher - monatliche Daten"                                                                                 
#>  [4] "ei_bsco_m"                                                                                                      
#>  [5] "29.09.2021"                                                                                                     
#>  [6] "29.09.2021"                                                                                                     
#>  [7] "1980M01"                                                                                                        
#>  [8] "2021M09"                                                                                                        
#>  [9] "284580"                                                                                                         
#> [10] ""                                                                                                               
#> [11] ""                                                                                                               
#> [12] ""                                                                                                               
#> [13] ""                                                                                                               
#> [14] ""                                                                                                               
#> [15] ""                                                                                                               
#> [16] "https://ec.europa.eu/eurostat/cache/metadata/en/ei_bcs_esms.htm"                                                
#> [17] "https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bcs_esms.sdmx.zip"
#> [18] "https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bsco_m.tsv.gz"        
#> [19] "https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bsco_m.sdmx.zip"      

options(timeout=60)
# }