Download the Data Structure Definition of a dataset

Download Data Structure Definition (DSD) of a Eurostat dataset if it is not cached previously.

get_eurostat_dsd(
  id,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  compress_file = TRUE,
  verbose = FALSE,
  ...
)

Arguments

id
a character string with the id of the dataset. It is the value from the codename column of the get_eurostat_toc function.

cache
a boolean whether to load/save the TOC from/in the cache or not. The default value is TRUE, so that the TOC is checked first in the cache and if does not exist then downloaded from Eurostat and cached.

update_cache
a boolean to update cache or not. The default value is FALSE, so the cache is not updated. Can be set also with options(restatapi_update=TRUE)

cache_dir
a path to a cache directory. The default is NULL, in this case the TOC is cached in the memory (in the '.restatapi_env'). Otherwise if the cache_dir directory does not exist it creates the 'restatapi' directory in the temporary directory from tempdir() to save the RDS-file. Directory can also be set with option(restatapi_cache_dir=...).

compress_file
a logical whether to compress the RDS-file in caching. Default is TRUE.

verbose
A boolean with default FALSE, so detailed messages (for debugging) will not printed. Can be set also with options(restatapi_verbose=TRUE)

...
parameter to pass on the load_cfg function

Value

If the DSD does not exist it returns NULL otherwise the result is a table with the 3 columns:

concept	The name of the concepts in the order of the data structure

code	The possible list of codes under the concept

name	The name/description of the code

Details

The DSD is downloaded from Eurostat's website, through the REST API in XML (SDMX) format.

References

For more information see the detailed documentation of the API.

See also

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}
# }
dsd<-get_eurostat_dsd("nama_10_gdp",cache=FALSE,verbose=TRUE)

#> Trying to download the DSD from: http://ec.europa.eu/eurostat/SDMX/diss-web/rest/datastructure/ESTAT/DSD_nama_10_gdp

#> Trying to extract the DSD from: /tmp/RtmpH49KR8/file1a62abafe65

head(dsd)

#>   concept code        name
#> 1    FREQ    D       Daily
#> 2    FREQ    W      Weekly
#> 3    FREQ    Q   Quarterly
#> 4    FREQ    A      Annual
#> 5    FREQ    S Semi-annual
#> 6    FREQ    M     Monthly