Download the Table of Contents of Eurostat datasets

Download Table of Contents (TOC) of Eurostat datasets if it is not cached previously.

get_eurostat_toc(
  mode = "xml",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  compress_file = TRUE,
  lang = "en",
  verbose = FALSE,
  ...
)

Arguments

mode

a character string either xml or txt defining the download mode. Depending on the mode the 'xml' version or the 'text' version of the TOC is downloaded. The deafult value is xml as it provides more information (e.g. number of values, short description and download links in different formats (SDMX, TSV))

cache

a boolean whether to load/save the TOC from/in the cache or not. The default value is TRUE, so that the TOC is checked first in the cache and if does not exist then downloaded from Eurostat and cached.

update_cache

a boolean to update cache or not. The default value is FALSE, so the cache is not updated. Can be set also with options(restatapi_update=TRUE)

cache_dir

a path to a cache directory. The default is NULL, in this case the TOC is cached in the memory (in the '.restatapi_env'). Otherwise if the cache_dir directory does not exist it creates the 'restatapi' directory in the temporary directory from tempdir() to save the RDS- file. Directory can also be set with option(restatapi_cache_dir=...).

compress_file

a logical whether to compress the RDS-file in caching. Default is TRUE.

lang

a character string either en, de or fr to define the language version for the table of contents. The default is en - English.

verbose

A boolean with default FALSE, so detailed messages (for debugging) will not printed. Can be set also with options(restatapi_verbose=TRUE)

...

parameter to pass on the load_cfg function

Value

A data table with the following columns:

title	The name of dataset/table in the language provided by the lang parameter

code	The codename of dataset/table which can be used as id in other functions

type	The type of information: 'dataset' or 'table'

lastUpdate	The date when the data was last time updated for tables and datasets

lastModified	The date when the structure of the dataset/table was last time modified

dataStart	The start date of the data in the dataset/table

dataEnd	The end date of the data in the dataset/table

values	The number of values in the dataset/table, and it is filled only if the download mode is "xml"

unit	The unit name for tables in the language provided by the lang parameter, for dataset it is empty and this column exists only if the download mode is "xml"

shortDescription	The short description of the values for tables in the language provided by the lang parameter, for dataset it is empty and this column exists only if the download mode is "xml"

metadata.html	The link to the metadata in html format, and this column exists only if the download mode is "xml"

metadata.sdmx	The link to the metadata in SDMX format, and this column exists only if the download mode is "xml"

downloadLink.tsv	The link to the whole dataset/table in tab separated values format in the bulk download facility and this column exists only if the download mode is "xml"

downloadLink.sdmx	The link to the whole dataset/table in SDMX format in the bulk download facility and this column exists only if the download mode is "xml"

Details

The TOC is downloaded from Eurostat websites through the REST API for the xml (default) version or from the bulk download facilities for txt version. From the downloaded TOC the values in the 'code' column can be used as id in the get_eurostat_dsd, get_eurostat_raw, get_eurostat_bulk, and get_eurostat_data functions.

References

For more technical information see the detailed documentation of the API.

See also

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}    
# }
# \donttest{
options(timeout=2)
toc_xml<-get_eurostat_toc(cache=FALSE,verbose=TRUE)
#> 
#> get_eurostat_toc - API version:2 - number of cores:1
#> get_eurostat_toc - API version:2 - number of cores:1
#> get_eurostat_toc - Downloading https://ec.europa.eu/eurostat/api/dissemination/catalogue/toc/xml
#> xml_nodeset
#> number of nodes: 8967
#> number of cores: 1
head(toc_xml)
#>                                                 title         code    type
#> 1:                   Current account - quarterly data  ei_bpm6ca_q   table
#> 2:                 Financial account - quarterly data  ei_bpm6fa_q   table
#> 3:                     Current account - monthly data  ei_bpm6ca_m   table
#> 4:                   Financial account - monthly data  ei_bpm6fa_m   table
#> 5: International investment position - quarterly data ei_bpm6iip_q   table
#> 6:                           Consumers - monthly data    ei_bsco_m dataset
#>    lastUpdate lastModified dataStart dataEnd values unit shortDescription
#> 1: 2023.04.24   2023.04.24   1992-Q1 2022-Q4 264378                      
#> 2: 2023.04.24   2023.04.24   1992-Q1 2022-Q4  47735                      
#> 3: 2023.04.19   2023.04.19   1991-01 2023-02 210914                      
#> 4: 2023.04.19   2023.04.19   1991-01 2023-02  69171                      
#> 5: 2023.04.25   2023.04.25   1992-Q4 2022-Q4  57671                      
#> 6: 2023.04.27   2023.04.27   1980-01 2023-04 300645                      
#>                                                      metadata.html
#> 1:  https://ec.europa.eu/eurostat/cache/metadata/en/ei_bp_esms.htm
#> 2:  https://ec.europa.eu/eurostat/cache/metadata/en/ei_bp_esms.htm
#> 3:  https://ec.europa.eu/eurostat/cache/metadata/en/ei_bp_esms.htm
#> 4:  https://ec.europa.eu/eurostat/cache/metadata/en/ei_bp_esms.htm
#> 5:  https://ec.europa.eu/eurostat/cache/metadata/en/ei_bp_esms.htm
#> 6: https://ec.europa.eu/eurostat/cache/metadata/en/ei_bcs_esms.htm
#>                                                                                                      metadata.sdmx
#> 1:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bp_esms.sdmx.zip
#> 2:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bp_esms.sdmx.zip
#> 3:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bp_esms.sdmx.zip
#> 4:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bp_esms.sdmx.zip
#> 5:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bp_esms.sdmx.zip
#> 6: https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_bcs_esms.sdmx.zip
#>                                                                                              downloadLink.tsv
#> 1:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6ca_q.tsv.gz
#> 2:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6fa_q.tsv.gz
#> 3:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6ca_m.tsv.gz
#> 4:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6fa_m.tsv.gz
#> 5: https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6iip_q.tsv.gz
#> 6:    https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bsco_m.tsv.gz
#>                                                                                               downloadLink.sdmx
#> 1:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6ca_q.sdmx.zip
#> 2:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6fa_q.sdmx.zip
#> 3:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6ca_m.sdmx.zip
#> 4:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6fa_m.sdmx.zip
#> 5: https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bpm6iip_q.sdmx.zip
#> 6:    https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_bsco_m.sdmx.zip
toc_txt<-get_eurostat_toc(mode="txt", lang="de")
head(toc_txt)
#> NULL
options(timeout=60)
# }