Load configuration data from JSON

Load the configuration information to the '.restatapi_env' from the JSON configuration file.

load_cfg(api_version = "current", load_toc = FALSE, parallel = TRUE,
  max_cores = TRUE, verbose = FALSE)

Arguments

api_version
It can be either "old", "new", "test" or "current". The default value is "current".

load_toc
The default value FALSE, which means that the Table of contents (TOC) will not be downloaded and cached automatically in the '.restatapi_env' when the package is loaded.

parallel
A boolean with the default value TRUE. If there are multiple cores/logical processors then part of the data extraction is made in parallel reducing significantly the time needed for large datasets. If the value is FALSE the option restatapi_cores set to 1.

max_cores
A boolean with the default value TRUE. If the parameter 'parallel' is TRUE then this parameter is taken into account otherwise it is ignored. If the default value is used, then the maximum minus one cores/logical processors are used for parallel computing. If the parameter FALSE, then the default value of getOption("mc.cores") is used, if it is defined. If mc.cores is NULL and there are more than 2 cores/logical processors then the option restatapi_cores are set to 2. Otherwise the parallel processing turned off by setting the option restatapi_cores to 1. The number of cores used for parallel computing can be changed any time with options(restatapi_cores=...)

verbose
A boolean if the verbose message about the configuration to be showed or not. The default is FALSE. Can be set also with options(restatapi_verbose=TRUE)

Value

it returns 4 objects in the '.restatapi_env'

cfg a list with all the configuration data

rav a character string with a number defining the API_VERSION from the config file to be used later. It is determined based on the api_version parameter.

cc a list containing the 2 character country codes of the member states for different EU composition like EU15, EU28 or EA (Euro Area).

dmethod the download method to be used to access Eurostat database. If the 'libcurl' method exists then it will be the method for file download otherwise it will be 'auto'.

Details

Loads configuration data from a JSON file. The function first tries to load the configuration file from GitHub. If it is not possible it loads from the file delivered with the package.By this way different version of the API can be tested. In addition, the list of country codes are loaded to the variable cc (country codes), based on the Eurostat standard code list

Examples

# \donttest{
load_cfg(parallel=FALSE)
load_cfg(api_version="test",verbose=TRUE,max_cores=FALSE)

#> restatapi: - config file with the API version 3 loaded from GitHub (the 'current' API version number is 1).
#>            - 2 from the 4 cores are used for parallel computing.
#>            - 'libcurl' will be used for file download.
#>            - the Table of contents (TOC) was not pre-loaded into the deafult cache ('.restatapi_env').

load_cfg()
eu<-get("cc",envir=.restatapi_env)
eu$EU28

#>  [1] "BE" "BG" "CZ" "DK" "DE" "EE" "IE" "EL" "ES" "FR" "HR" "IT" "CY" "LV" "LT"
#> [16] "LU" "HU" "MT" "NL" "AT" "PL" "PT" "RO" "SI" "SK" "FI" "SE" "UK"

eu$EA15

#>  [1] "BE" "DE" "IE" "EL" "ES" "FR" "IT" "CY" "LU" "MT" "NL" "AT" "PT" "SI" "FI"

# }