Search for pattern in the Data Structure Definition of a dataset

Search the Data Structure Definition (DSD) of a Eurostat dataset for a given pattern. It returns the rows where the pattern appears in the code and name column of the output of the get_eurostat_dsd function.

search_eurostat_dsd(pattern, dsd = NULL, name = TRUE, exact_match = FALSE, ...)

Arguments

pattern
a character string or a vector of character string.

dsd
a table containing Data Structure Definition (DSD) of a Eurostat dataset which can be retreived by the get_eurostat_dsd function.

name
a boolean with the default value TRUE, if the search shall look for the pattern in the name of the code. If the value FALSE, then only the 'code' column of the DSD will be searched.

exact_match
a boolean with the default value FALSE, if the strings provided in pattern shall be matched exactly as it is or as a pattern.

...
additional arguments to the grep function like ignore.case=TRUE if the pattern should be searched case sensitive or not. The default value for ignore.case is FALSE.

Value

If the pattern found then the function returns a data.frame with the 4 columns:

pattern	The pattern which was searched

concept	The name of the concepts in the data structure

code	The list of codes where the pattern was found, or the code of a name where the pattern appears

name	The name/description of the code where the pattern found, or the name of the code where the pattern appears

Otherwise returns the value NULL.

Details

The function returns the line(s) where the searched pattern appears in the code or in the name column.

See also

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}    
cfg<-get("cfg",envir=.restatapi_env) 
rav<-get("rav",envir=.restatapi_env)
# }
dsd_example<-get_eurostat_dsd("nama_10_gdp",verbose=TRUE)

#> Trying to download the DSD from: http://ec.europa.eu/eurostat/SDMX/diss-web/rest/datastructure/ESTAT/DSD_nama_10_gdp

#> Trying to extract the DSD from: /tmp/Rtmp0DocIa/file1a720673bc0

#> The DSD of the nama_10_gdp dataset was cached in memory ('nama_10_gdp.dsd' in '.restatapi_env').

search_eurostat_dsd("EU",dsd_example)

#>    pattern concept                 code
#> 12      EU    UNIT PC_EU27_2020_MEUR_CP
#> 13      EU    UNIT      PC_EU28_MEUR_CP
#> 14      EU    UNIT PC_EU27_2020_MPPS_CP
#> 15      EU    UNIT      PC_EU28_MPPS_CP
#> 16      EU    UNIT              CP_MEUR
#> 19      EU    UNIT    CP_MPPS_EU27_2020
#> 20      EU    UNIT           CLV15_MEUR
#> 21      EU    UNIT           CLV10_MEUR
#> 22      EU    UNIT           CLV05_MEUR
#> 27      EU    UNIT             PYP_MEUR
#> 30      EU    UNIT             PD15_EUR
#> 31      EU    UNIT             PD10_EUR
#> 32      EU    UNIT             PD05_EUR
#> 36      EU    UNIT       PD_PCH_PRE_EUR
#> 77      EU     GEO            EU27_2020
#> 78      EU     GEO                 EU28
#> 79      EU     GEO                 EU15
#>                                                                                                  name
#> 12                       Percentage of EU27 (from 2020) total (based on million euro), current prices
#> 13                       Percentage of EU28 (2013-2020) total (based on million euro), current prices
#> 14 Percentage of EU27 (from 2020) total (based on million purchasing power standards), current prices
#> 15 Percentage of EU28 (2013-2020) total (based on million purchasing power standards), current prices
#> 16                                                                       Current prices, million euro
#> 19                           Current prices, million purchasing power standards (PPS, EU27 from 2020)
#> 20                                                          Chain linked volumes (2015), million euro
#> 21                                                          Chain linked volumes (2010), million euro
#> 22                                                          Chain linked volumes (2005), million euro
#> 27                                                                 Previous year prices, million euro
#> 30                                                    Price index (implicit deflator), 2015=100, euro
#> 31                                                    Price index (implicit deflator), 2010=100, euro
#> 32                                                    Price index (implicit deflator), 2005=100, euro
#> 36                        Price index (implicit deflator), percentage change on previous period, euro
#> 77                                                          European Union - 27 countries (from 2020)
#> 78                                                          European Union - 28 countries (2013-2020)
#> 79                                                          European Union - 15 countries (1995-2004)

search_eurostat_dsd("EU",dsd_example,ignore.case=TRUE)

#>     pattern  concept                 code
#> 12       EU     UNIT PC_EU27_2020_MEUR_CP
#> 13       EU     UNIT      PC_EU28_MEUR_CP
#> 14       EU     UNIT PC_EU27_2020_MPPS_CP
#> 15       EU     UNIT      PC_EU28_MPPS_CP
#> 16       EU     UNIT              CP_MEUR
#> 19       EU     UNIT    CP_MPPS_EU27_2020
#> 20       EU     UNIT           CLV15_MEUR
#> 21       EU     UNIT           CLV10_MEUR
#> 22       EU     UNIT           CLV05_MEUR
#> 27       EU     UNIT             PYP_MEUR
#> 30       EU     UNIT             PD15_EUR
#> 31       EU     UNIT             PD10_EUR
#> 32       EU     UNIT             PD05_EUR
#> 36       EU     UNIT       PD_PCH_PRE_EUR
#> 77       EU      GEO            EU27_2020
#> 78       EU      GEO                 EU28
#> 79       EU      GEO                 EU15
#> 80       EU      GEO                   EA
#> 81       EU      GEO                 EA19
#> 82       EU      GEO                 EA12
#> 126      EU OBS_FLAG                    s
#>                                                                                                   name
#> 12                        Percentage of EU27 (from 2020) total (based on million euro), current prices
#> 13                        Percentage of EU28 (2013-2020) total (based on million euro), current prices
#> 14  Percentage of EU27 (from 2020) total (based on million purchasing power standards), current prices
#> 15  Percentage of EU28 (2013-2020) total (based on million purchasing power standards), current prices
#> 16                                                                        Current prices, million euro
#> 19                            Current prices, million purchasing power standards (PPS, EU27 from 2020)
#> 20                                                           Chain linked volumes (2015), million euro
#> 21                                                           Chain linked volumes (2010), million euro
#> 22                                                           Chain linked volumes (2005), million euro
#> 27                                                                  Previous year prices, million euro
#> 30                                                     Price index (implicit deflator), 2015=100, euro
#> 31                                                     Price index (implicit deflator), 2010=100, euro
#> 32                                                     Price index (implicit deflator), 2005=100, euro
#> 36                         Price index (implicit deflator), percentage change on previous period, euro
#> 77                                                           European Union - 27 countries (from 2020)
#> 78                                                           European Union - 28 countries (2013-2020)
#> 79                                                           European Union - 15 countries (1995-2004)
#> 80  Euro area (EA11-1999, EA12-2001, EA13-2007, EA15-2008, EA16-2009, EA17-2011, EA18-2014, EA19-2015)
#> 81                                                               Euro area - 19 countries  (from 2015)
#> 82                                                                Euro area - 12 countries (2001-2006)
#> 126                                                                     Eurostat estimate (phased out)

search_eurostat_dsd("EU27_2019",dsd_example,name=FALSE)

#> NULL

search_eurostat_dsd("EU27_2019",dsd_example,exact_match=TRUE)

#> NULL