Search for pattern in the titles, units and short description of the TOC

Lists names of dataset from Eurostat with the particular pattern in the title, units or short description.

search_eurostat_toc(pattern, lang = "en", verbose = FALSE, ...)

Arguments

pattern
Character string to search for in the table of contents of Eurostat tables/datasets

lang
a character string either en, de or fr to define the language version for the table of contents. The default is en - English.

verbose
A boolean with default FALSE, so detailed messages (for debugging) will not printed. Can be set also with options(restatapi_verbose=TRUE)

...
other additional parameters to pass to the grepl function like ignore.case=TRUE if the pattern should be searched case sensitive or not. The default value for ignore.case is FALSE.

Value

A table with the following columns:

title	The name of dataset/table in the language provided by the lang parameter

code	The codename of dataset/table which can be used by the get_eurostat function

type	The type of information: 'dataset' or 'table'

lastUpdate	The date when the data was last time updated for tables and datasets

lastModified	The date when the structure of the dataset/table was last time modified

dataStart	The start date of the data in the dataset/table

dataEnd	The end date of the data in the dataset/table

values	The number of values in the dataset/table

unit	The unit name for tables in the language provided by the lang parameter, if the type 'dataset' this column is empty

shortDescription	The short description of the values for tables in the language provided by the lang parameterif the type 'dataset' this column is empty

metadata.html	The link to the metadata in html format

metadata.sdmx	The link to the metadata in SDMX format

downloadLink.tsv	The link to the whole dataset/table in tab separated values format in the bulk download facility

downloadLink.sdmx	The link to the whole dataset/table in SDMX format in the bulk download facility

The value in the code column can be used as an id in the get_eurostat_data, get_eurostat_bulk, get_eurostat_raw and get_eurostat_dsd functions. If there is no hit for the search query, it returns NULL.

Details

Downloads the list of all tables and datasets available in the Eurostat database and returns all the details from the table of contents of the tables/datasets that contains particular pattern in the dataset title, unit or short description. E.g. all tables/datasets mentioning 'energy'.

See also

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}    
# }
# \donttest{
options(timeout=2)
head(search_eurostat_toc("energy",verbose=TRUE))
#> get_eurostat_toc - Downloading https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&downfile=table_of_contents.xml
#> xml_nodeset
#>  number of nodes: 8736
#> get_eurostat_toc - The TOC was cached in memory ('toc.xml.en' in '.restatapi_env').
#>                                                                                                                  title
#> 699                                                                                 Final energy consumption by sector
#> 726                                                               Candidate countries and potential candidates: energy
#> 879                                                                                          Primary energy production
#> 880                                                                       Secondary energy production (transformation)
#> 907                                                                                              ENP countries: energy
#> 1496 Psychological well-being (energy and vitality index) by sex, age and educational attainment level (average score)
#>                 code    type lastUpdate lastModified dataStart dataEnd values
#> 699  enps_nrg_bal_c1 dataset 2021.04.16         <NA>      2005    2019    267
#> 726       cpc_energy dataset 2016.10.11   2021.02.08      2000    2015    926
#> 879         med_eg10 dataset 2020.05.28   2021.02.08      2005    2018    391
#> 880         med_eg20 dataset 2016.10.28   2021.02.08      2000    2015    776
#> 907      enpr_energy dataset 2020.01.14   2021.02.08      2005    2018    262
#> 1496   hlth_ehis_st8 dataset 2014.03.20   2021.02.08      2008    2008   1455
#>      unit shortDescription
#> 699                       
#> 726                       
#> 879                       
#> 880                       
#> 907                       
#> 1496                      
#>                                                          metadata.html
#> 699      https://ec.europa.eu/eurostat/cache/metadata/en/enps_esms.htm
#> 726       https://ec.europa.eu/eurostat/cache/metadata/en/cpc_esms.htm
#> 879       https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 880       https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 907      https://ec.europa.eu/eurostat/cache/metadata/en/enpr_esms.htm
#> 1496 https://ec.europa.eu/eurostat/cache/metadata/en/hlth_det_esms.htm
#>                                                                                                          metadata.sdmx
#> 699      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/enps_esms.sdmx.zip
#> 726       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/cpc_esms.sdmx.zip
#> 879       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 880       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 907      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/enpr_esms.sdmx.zip
#> 1496 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/hlth_det_esms.sdmx.zip
#>                                                                                                   downloadLink.tsv
#> 699  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.tsv.gz
#> 726       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.tsv.gz
#> 879         https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 880         https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#> 907      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.tsv.gz
#> 1496   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.tsv.gz
#>                                                                                                    downloadLink.sdmx
#> 699  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.sdmx.zip
#> 726       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.sdmx.zip
#> 879         https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 880         https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
#> 907      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.sdmx.zip
#> 1496   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.sdmx.zip
nrow(search_eurostat_toc("energy"))
#> [1] 96
head(search_eurostat_toc("energie",lang="de",ignore.case=TRUE))
#>                                                     title            code
#> 38                             Energie - monatliche Daten       ei_isen_m
#> 698                                         Energiebilanz  enps_nrg_bal_c
#> 699                       Endenergieverbrauch nach Sektor enps_nrg_bal_c1
#> 726 Kandidaten- und potentielle Kandidatenländer: Energie      cpc_energy
#> 879                               Primäre Energierzeugung        med_eg10
#> 880                 Sekundäre Energierzeugung (Umsetzung)        med_eg20
#>        type lastUpdate lastModified dataStart dataEnd values unit
#> 38  dataset 2021.11.18   2021.11.18   2008M01 2021M10  73265     
#> 698 dataset 2021.04.16         <NA>      2005    2019    518     
#> 699 dataset 2021.04.16         <NA>      2005    2019    267     
#> 726 dataset 2016.10.11   2021.02.08      2000    2015    926     
#> 879 dataset 2020.05.28   2021.02.08      2005    2018    391     
#> 880 dataset 2016.10.28   2021.02.08      2000    2015    776     
#>     shortDescription
#> 38                  
#> 698                 
#> 699                 
#> 726                 
#> 879                 
#> 880                 
#>                                                      metadata.html
#> 38  https://ec.europa.eu/eurostat/cache/metadata/en/ei_is_esms.htm
#> 698  https://ec.europa.eu/eurostat/cache/metadata/en/enps_esms.htm
#> 699  https://ec.europa.eu/eurostat/cache/metadata/en/enps_esms.htm
#> 726   https://ec.europa.eu/eurostat/cache/metadata/en/cpc_esms.htm
#> 879   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 880   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#>                                                                                                      metadata.sdmx
#> 38  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_is_esms.sdmx.zip
#> 698  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/enps_esms.sdmx.zip
#> 699  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/enps_esms.sdmx.zip
#> 726   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/cpc_esms.sdmx.zip
#> 879   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 880   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#>                                                                                                  downloadLink.tsv
#> 38        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.tsv.gz
#> 698  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c.tsv.gz
#> 699 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.tsv.gz
#> 726      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.tsv.gz
#> 879        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 880        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#>                                                                                                   downloadLink.sdmx
#> 38        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.sdmx.zip
#> 698  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c.sdmx.zip
#> 699 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.sdmx.zip
#> 726      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.sdmx.zip
#> 879        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 880        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
nrow(search_eurostat_toc("energie",lang="de",ignore.case=TRUE))
#> [1] 111
options(timeout=60)
# }