Search for pattern in the titles, units and short description of the TOC

Lists names of dataset from Eurostat with the particular pattern in the title, units or short description.

search_eurostat_toc(pattern, lang = "en", verbose = FALSE, ...)

Arguments

pattern

Character string to search for in the table of contents of Eurostat tables/datasets

lang

a character string either en, de or fr to define the language version for the table of contents. The default is en - English.

verbose

A boolean with default FALSE, so detailed messages (for debugging) will not printed. Can be set also with options(restatapi_verbose=TRUE)

...

other additional parameters to pass to the grepl function like ignore.case=TRUE if the pattern should be searched case sensitive or not. The default value for ignore.case is FALSE.

Value

A table with the following columns:

title	The name of dataset/table in the language provided by the lang parameter

code	The codename of dataset/table which can be used by the get_eurostat function

type	The type of information: 'dataset' or 'table'

lastUpdate	The date when the data was last time updated for tables and datasets

lastModified	The date when the structure of the dataset/table was last time modified

dataStart	The start date of the data in the dataset/table

dataEnd	The end date of the data in the dataset/table

values	The number of values in the dataset/table

unit	The unit name for tables in the language provided by the lang parameter, if the type 'dataset' this column is empty

shortDescription	The short description of the values for tables in the language provided by the lang parameterif the type 'dataset' this column is empty

metadata.html	The link to the metadata in html format

metadata.sdmx	The link to the metadata in SDMX format

downloadLink.tsv	The link to the whole dataset/table in tab separated values format in the bulk download facility

downloadLink.sdmx	The link to the whole dataset/table in SDMX format in the bulk download facility

The value in the code column can be used as an id in the get_eurostat_data, get_eurostat_bulk, get_eurostat_raw and get_eurostat_dsd functions. If there is no hit for the search query, it returns NULL.

Details

Downloads the list of all tables and datasets available in the Eurostat database and returns all the details from the table of contents of the tables/datasets that contains particular pattern in the dataset title, unit or short description. E.g. all tables/datasets mentioning 'energy'.

See also

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}    
# }
# \donttest{
options(timeout=2)
head(search_eurostat_toc("energy",verbose=TRUE))
#> 
#> get_eurostat_toc - API version:2 - number of cores:2
#> get_eurostat_toc - Downloading https://ec.europa.eu/eurostat/api/dissemination/catalogue/toc/xml
#> get_eurostat_toc - class(xml_leafs): xml_nodeset
#> get_eurostat_toc - number of nodes: 9246
#> get_eurostat_toc - number of cores: 2
#> get_eurostat_toc - The TOC was cached in memory ('toc.xml.en' in '.restatapi_env').
#>                                                                                                                title
#> 1:                                                                                Final energy consumption by sector
#> 2:                                                                                         Primary energy production
#> 3:                                                                      Secondary energy production (transformation)
#> 4:                                                              Candidate countries and potential candidates: energy
#> 5:                                                                                             ENP countries: energy
#> 6: Psychological well-being (energy and vitality index) by sex, age and educational attainment level (average score)
#>               code    type lastUpdate lastModified dataStart dataEnd values
#> 1: enps_nrg_bal_c1 dataset 2023.03.13   2023.03.13      2005    2021    305
#> 2:        med_eg10 dataset 2020.05.28   2020.05.28      2005    2018    391
#> 3:        med_eg20 dataset 2016.10.28   2023.12.21      2000    2015    776
#> 4:      cpc_energy dataset 2016.10.11   2023.12.21      2000    2015    926
#> 5:     enpr_energy dataset 2020.01.14   2020.01.14      2005    2018    262
#> 6:   hlth_ehis_st8 dataset 2014.03.20   2023.12.21      2008    2008   1455
#>    unit shortDescription
#> 1:                      
#> 2:                      
#> 3:                      
#> 4:                      
#> 5:                      
#> 6:                      
#>                                                        metadata.html
#> 1:     https://ec.europa.eu/eurostat/cache/metadata/en/enps_esms.htm
#> 2:      https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 3:      https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 4:      https://ec.europa.eu/eurostat/cache/metadata/en/cpc_esms.htm
#> 5:     https://ec.europa.eu/eurostat/cache/metadata/en/enpr_esms.htm
#> 6: https://ec.europa.eu/eurostat/cache/metadata/en/hlth_det_esms.htm
#>                                                                                 metadata.sdmx
#> 1:     https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/enps_esms.sdmx.zip
#> 2:      https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/med_esms.sdmx.zip
#> 3:      https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/med_esms.sdmx.zip
#> 4:      https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/cpc_esms.sdmx.zip
#> 5:     https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/enpr_esms.sdmx.zip
#> 6: https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/hlth_det_esms.sdmx.zip
#>                                                                                                 downloadLink.tsv
#> 1: https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.tsv.gz
#> 2:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 3:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#> 4:      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.tsv.gz
#> 5:     https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.tsv.gz
#> 6:   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.tsv.gz
#>                                                                                                  downloadLink.sdmx
#> 1: https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.sdmx.zip
#> 2:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 3:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
#> 4:      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.sdmx.zip
#> 5:     https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.sdmx.zip
#> 6:   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.sdmx.zip
nrow(search_eurostat_toc("energy"))
#> [1] 94
head(search_eurostat_toc("energie",lang="de",ignore.case=TRUE))
#>                                                                    title
#> 1:                                            Energie - monatliche Daten
#> 2:                                                         Energiebilanz
#> 3:                                       Endenergieverbrauch nach Sektor
#> 4: Energie - historische daten (2004-2007, ehemalige Methode und Quelle)
#> 5:                                               Primäre Energierzeugung
#> 6:                                 Sekundäre Energierzeugung (Umsetzung)
#>               code    type lastUpdate lastModified dataStart dataEnd values
#> 1:       ei_isen_m   table 2024.01.06   2024.01.06   2008-01 2023-11  72517
#> 2:  enps_nrg_bal_c dataset 2023.03.13   2023.03.13      2005    2021    644
#> 3: enps_nrg_bal_c1 dataset 2023.03.13   2023.03.13      2005    2021    305
#> 4:      med_eg_all dataset 2014.12.08   2023.12.21      1990    2011    935
#> 5:        med_eg10 dataset 2020.05.28   2020.05.28      2005    2018    391
#> 6:        med_eg20 dataset 2016.10.28   2023.12.21      2000    2015    776
#>    unit shortDescription
#> 1:                      
#> 2:                      
#> 3:                      
#> 4:                      
#> 5:                      
#> 6:                      
#>                                                     metadata.html
#> 1: https://ec.europa.eu/eurostat/cache/metadata/en/ei_is_esms.htm
#> 2:  https://ec.europa.eu/eurostat/cache/metadata/en/enps_esms.htm
#> 3:  https://ec.europa.eu/eurostat/cache/metadata/en/enps_esms.htm
#> 4:   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 5:   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 6:   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#>                                                                              metadata.sdmx
#> 1: https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/ei_is_esms.sdmx.zip
#> 2:  https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/enps_esms.sdmx.zip
#> 3:  https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/enps_esms.sdmx.zip
#> 4:   https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/med_esms.sdmx.zip
#> 5:   https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/med_esms.sdmx.zip
#> 6:   https://ec.europa.eu/eurostat/api/dissemination/files?file=metadata/med_esms.sdmx.zip
#>                                                                                                 downloadLink.tsv
#> 1:       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.tsv.gz
#> 2:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c.tsv.gz
#> 3: https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.tsv.gz
#> 4:      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg_all.tsv.gz
#> 5:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 6:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#>                                                                                                  downloadLink.sdmx
#> 1:       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.sdmx.zip
#> 2:  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c.sdmx.zip
#> 3: https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enps_nrg_bal_c1.sdmx.zip
#> 4:      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg_all.sdmx.zip
#> 5:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 6:        https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
nrow(search_eurostat_toc("energie",lang="de",ignore.case=TRUE))
#> [1] 107
options(timeout=60)
# }