Search in Eurostat datasets titles, units and short description

Lists names of dataset from Eurostat with the particular pattern in the title, units or short description.

search_eurostat_toc(pattern, lang = "en", verbose = FALSE, ...)

Arguments

pattern
Character string to search for in the table of contents of Eurostat tables/datasets

lang
a character string either en, de or fr to define the language version for the table of contents. The default is en - English.

verbose
A boolean with default FALSE, so detailed messages (for debugging) will not printed. Can be set also with options(restatapi_verbose=TRUE)

...
other parameters to pass to the grep function.

Value

A table with the following columns:

title	The name of dataset/table in the language provided by the lang parameter

code	The codename of dataset/table which can be used by the get_eurostat function

type	The type of information: 'dataset' or 'table'

lastUpdate	The date when the data was last time updated for tables and datasets

lastModified	The date when the structure of the dataset/table was last time modified

dataStart	The start date of the data in the dataset/table

dataEnd	The end date of the data in the dataset/table

values	The number of values in the dataset/table.

unit	The unit name for tables in the language provided by the lang parameter. For dataset it is empty.

shortDescription	The short description of the values for tables in the language provided by the lang parameter. For dataset it is empty.

metadata.html	The link to the metadata in html format.

metadata.sdmx	The link to the metadata in SDMX format.

downloadLink.tsv	The link to the whole dataset/table in tab separated values format in the bulk download facility.

downloadLink.sdmx	The link to the whole dataset/table in SDMX format in the bulk download facility.

The value in the code column can be used as an id in the get_eurostat_data, get_eurostat_bulk, get_eurostat_raw and get_eurostat_dsd functions.

Details

Downloads the list of all tables and datasets available in the Eurostat database and returns all the details from the table of contents of the tables/datasets that contains particular pattern in the dataset title, unit or short description. E.g. all tables/datasets mentioning 'energy'.

See also

Examples

  head(search_eurostat_toc("energy"))

#>                                                                                                                  title
#> 664                                                                                          Primary energy production
#> 665                                                                       Secondary energy production (transformation)
#> 700                                                               Candidate countries and potential candidates: energy
#> 733                                                                                              ENP countries: energy
#> 1273 Psychological well-being (energy and vitality index) by sex, age and educational attainment level (average score)
#> 3894             Multi-yearly enterprise statistics - purchases of energy products (NACE Rev. 1.1, C, D, F, 1997-2001)
#>               code    type lastUpdate lastModified dataStart dataEnd values
#> 664       med_eg10 dataset 2020-02-06   2020-01-04      2005    2017    367
#> 665       med_eg20 dataset 2020-10-28   2020-05-07      2000    2015    776
#> 700     cpc_energy dataset 2020-10-11   2020-05-18      2000    2015    926
#> 733    enpr_energy dataset 2020-01-04   2020-01-04      2005    2017    239
#> 1273 hlth_ehis_st8 dataset 2020-03-20   2020-12-11      2008    2008   1455
#> 3894  sbs_pu_2l_97 dataset 2020-03-26   2020-06-03      1997    2001  77962
#>      unit shortDescription
#> 664                       
#> 665                       
#> 700                       
#> 733                       
#> 1273                      
#> 3894                      
#>                                                          metadata.html
#> 664       https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 665       https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 700       https://ec.europa.eu/eurostat/cache/metadata/en/cpc_esms.htm
#> 733      https://ec.europa.eu/eurostat/cache/metadata/en/enpr_esms.htm
#> 1273 https://ec.europa.eu/eurostat/cache/metadata/en/hlth_det_esms.htm
#> 3894   https://ec.europa.eu/eurostat/cache/metadata/en/sbs_pu_esms.htm
#>                                                                                                          metadata.sdmx
#> 664       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 665       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 700       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/cpc_esms.sdmx.zip
#> 733      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/enpr_esms.sdmx.zip
#> 1273 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/hlth_det_esms.sdmx.zip
#> 3894   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/sbs_pu_esms.sdmx.zip
#>                                                                                                downloadLink.tsv
#> 664       http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 665       http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#> 700     http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.tsv.gz
#> 733    http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.tsv.gz
#> 1273 http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.tsv.gz
#> 3894  http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/sbs_pu_2l_97.tsv.gz
#>                                                                                                 downloadLink.sdmx
#> 664       http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 665       http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
#> 700     http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.sdmx.zip
#> 733    http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.sdmx.zip
#> 1273 http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.sdmx.zip
#> 3894  http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/sbs_pu_2l_97.sdmx.zip

  head(search_eurostat_toc("energie",lang="de",ignore.case=TRUE))

#>                                                                     title
#> 35                                             Energie - monatliche Daten
#> 664                                               Primäre Energierzeugung
#> 665                                 Sekundäre Energierzeugung (Umsetzung)
#> 666                                                      Energieverbrauch
#> 667 Energie - historische daten (2004-2007, ehemalige Methode und Quelle)
#> 700                 Kandidaten- und potentielle Kandidatenländer: Energie
#>           code    type lastUpdate lastModified dataStart dataEnd values unit
#> 35   ei_isen_m dataset 2020-07-19   2020-07-19   2008M01 2019M06  43587     
#> 664   med_eg10 dataset 2020-02-06   2020-01-04      2005    2017    367     
#> 665   med_eg20 dataset 2020-10-28   2020-05-07      2000    2015    776     
#> 666   med_eg30 dataset 2020-02-06   2020-01-04      2005    2017    164     
#> 667 med_eg_all dataset 2020-12-08   2020-12-08      1990    2011    935     
#> 700 cpc_energy dataset 2020-10-11   2020-05-18      2000    2015    926     
#>     shortDescription
#> 35                  
#> 664                 
#> 665                 
#> 666                 
#> 667                 
#> 700                 
#>                                                      metadata.html
#> 35  https://ec.europa.eu/eurostat/cache/metadata/en/ei_is_esms.htm
#> 664   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 665   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 666   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 667   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 700   https://ec.europa.eu/eurostat/cache/metadata/en/cpc_esms.htm
#>                                                                                                      metadata.sdmx
#> 35  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_is_esms.sdmx.zip
#> 664   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 665   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 666   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 667   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 700   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/cpc_esms.sdmx.zip
#>                                                                                            downloadLink.tsv
#> 35   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.tsv.gz
#> 664   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 665   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#> 666   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg30.tsv.gz
#> 667 http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg_all.tsv.gz
#> 700 http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.tsv.gz
#>                                                                                             downloadLink.sdmx
#> 35   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.sdmx.zip
#> 664   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 665   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
#> 666   http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg30.sdmx.zip
#> 667 http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg_all.sdmx.zip
#> 700 http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.sdmx.zip