Search in Eurostat datasets titles, units and short description

Lists names of dataset from Eurostat with the particular pattern in the title, units or short description.

search_eurostat_toc(pattern, lang = "en", verbose = FALSE, ...)

Arguments

pattern
Character string to search for in the table of contents of Eurostat tables/datasets

lang
a character string either en, de or fr to define the language version for the table of contents. The default is en - English.

verbose
A boolean with default FALSE, so detailed messages (for debugging) will not printed. Can be set also with options(restatapi_verbose=TRUE)

...
other additional parameters to pass to the grepl function like ignore.case=TRUE if the pattern should be searched case sensitive or not. The default value for ignore.case is FALSE.

Value

A table with the following columns:

title	The name of dataset/table in the language provided by the lang parameter

code	The codename of dataset/table which can be used by the get_eurostat function

type	The type of information: 'dataset' or 'table'

lastUpdate	The date when the data was last time updated for tables and datasets

lastModified	The date when the structure of the dataset/table was last time modified

dataStart	The start date of the data in the dataset/table

dataEnd	The end date of the data in the dataset/table

values	The number of values in the dataset/table

unit	The unit name for tables in the language provided by the lang parameter, if the type 'dataset' this column is empty

shortDescription	The short description of the values for tables in the language provided by the lang parameterif the type 'dataset' this column is empty

metadata.html	The link to the metadata in html format

metadata.sdmx	The link to the metadata in SDMX format

downloadLink.tsv	The link to the whole dataset/table in tab separated values format in the bulk download facility

downloadLink.sdmx	The link to the whole dataset/table in SDMX format in the bulk download facility

The value in the code column can be used as an id in the get_eurostat_data, get_eurostat_bulk, get_eurostat_raw and get_eurostat_dsd functions. If there is no hit for the search query it returns NULL.

Details

Downloads the list of all tables and datasets available in the Eurostat database and returns all the details from the table of contents of the tables/datasets that contains particular pattern in the dataset title, unit or short description. E.g. all tables/datasets mentioning 'energy'.

See also

Examples

# \dontshow{
if (parallel::detectCores()<=2){
   options(restatapi_cores=1)
}else{
   options(restatapi_cores=2)
}
# }
# \donttest{
  head(search_eurostat_toc("energy",verbose=TRUE))

#> Downloading https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&downfile=table_of_contents.xml

#> The TOC was cached in memory ('toc.xml.en' in '.restatapi_env').

#>                                                                                                                  title
#> 703                                                                                          Primary energy production
#> 704                                                                       Secondary energy production (transformation)
#> 733                                                               Candidate countries and potential candidates: energy
#> 760                                                                                              ENP countries: energy
#> 1301 Psychological well-being (energy and vitality index) by sex, age and educational attainment level (average score)
#> 3968             Multi-yearly enterprise statistics - purchases of energy products (NACE Rev. 1.1, C, D, F, 1997-2001)
#>               code    type lastUpdate lastModified dataStart dataEnd values
#> 703       med_eg10 dataset 2020.01.14   2020.02.27      2005    2018    387
#> 704       med_eg20 dataset 2016.10.28   2020.02.27      2000    2015    776
#> 733     cpc_energy dataset 2016.10.11   2020.02.27      2000    2015    926
#> 760    enpr_energy dataset 2020.01.14   2020.02.27      2005    2018    262
#> 1301 hlth_ehis_st8 dataset 2014.03.20   2020.02.27      2008    2008   1455
#> 3968  sbs_pu_2l_97 dataset 2009.03.26   2020.02.27      1997    2001  77962
#>      unit shortDescription
#> 703                       
#> 704                       
#> 733                       
#> 760                       
#> 1301                      
#> 3968                      
#>                                                          metadata.html
#> 703       https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 704       https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 733       https://ec.europa.eu/eurostat/cache/metadata/en/cpc_esms.htm
#> 760      https://ec.europa.eu/eurostat/cache/metadata/en/enpr_esms.htm
#> 1301 https://ec.europa.eu/eurostat/cache/metadata/en/hlth_det_esms.htm
#> 3968   https://ec.europa.eu/eurostat/cache/metadata/en/sbs_pu_esms.htm
#>                                                                                                          metadata.sdmx
#> 703       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 704       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 733       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/cpc_esms.sdmx.zip
#> 760      https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/enpr_esms.sdmx.zip
#> 1301 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/hlth_det_esms.sdmx.zip
#> 3968   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/sbs_pu_esms.sdmx.zip
#>                                                                                                 downloadLink.tsv
#> 703       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 704       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#> 733     https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.tsv.gz
#> 760    https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.tsv.gz
#> 1301 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.tsv.gz
#> 3968  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/sbs_pu_2l_97.tsv.gz
#>                                                                                                  downloadLink.sdmx
#> 703       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 704       https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
#> 733     https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.sdmx.zip
#> 760    https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/enpr_energy.sdmx.zip
#> 1301 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/hlth_ehis_st8.sdmx.zip
#> 3968  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/sbs_pu_2l_97.sdmx.zip

  nrow(search_eurostat_toc("energy"))

#> [1] 90

  head(search_eurostat_toc("energie",lang="de",ignore.case=TRUE))

#>                                                                     title
#> 36                                             Energie - monatliche Daten
#> 703                                               Primäre Energierzeugung
#> 704                                 Sekundäre Energierzeugung (Umsetzung)
#> 705                                                      Energieverbrauch
#> 706 Energie - historische daten (2004-2007, ehemalige Methode und Quelle)
#> 733                 Kandidaten- und potentielle Kandidatenländer: Energie
#>           code    type lastUpdate lastModified dataStart dataEnd values unit
#> 36   ei_isen_m dataset 2020.02.13   2020.02.27   2008M01 2019M12  46151     
#> 703   med_eg10 dataset 2020.01.14   2020.02.27      2005    2018    387     
#> 704   med_eg20 dataset 2016.10.28   2020.02.27      2000    2015    776     
#> 705   med_eg30 dataset 2020.01.14   2020.02.27      2005    2018    175     
#> 706 med_eg_all dataset 2014.12.08   2020.02.27      1990    2011    935     
#> 733 cpc_energy dataset 2016.10.11   2020.02.27      2000    2015    926     
#>     shortDescription
#> 36                  
#> 703                 
#> 704                 
#> 705                 
#> 706                 
#> 733                 
#>                                                      metadata.html
#> 36  https://ec.europa.eu/eurostat/cache/metadata/en/ei_is_esms.htm
#> 703   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 704   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 705   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 706   https://ec.europa.eu/eurostat/cache/metadata/en/med_esms.htm
#> 733   https://ec.europa.eu/eurostat/cache/metadata/en/cpc_esms.htm
#>                                                                                                      metadata.sdmx
#> 36  https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/ei_is_esms.sdmx.zip
#> 703   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 704   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 705   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 706   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/med_esms.sdmx.zip
#> 733   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=metadata/cpc_esms.sdmx.zip
#>                                                                                             downloadLink.tsv
#> 36   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.tsv.gz
#> 703   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.tsv.gz
#> 704   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.tsv.gz
#> 705   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg30.tsv.gz
#> 706 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg_all.tsv.gz
#> 733 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.tsv.gz
#>                                                                                              downloadLink.sdmx
#> 36   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/ei_isen_m.sdmx.zip
#> 703   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg10.sdmx.zip
#> 704   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg20.sdmx.zip
#> 705   https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg30.sdmx.zip
#> 706 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/med_eg_all.sdmx.zip
#> 733 https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=data/cpc_energy.sdmx.zip

  nrow(search_eurostat_toc("energie",lang="de",ignore.case=TRUE))

#> [1] 102

# }